feat: Updated User entity to autogenerate id
This commit is the result of about 40 minutes of debugging
with Ernie.
Previously, we get a ResultSet error when trying to create a
new User.
Set the id field to serial so that the spring boot and/or
hibernate side of things know to rely on the database to
generate the id.